fix: resolve CI failures — mypy, formatting, and test coverage
- Fix mypy: rename shadowed variable 'preds' to 'oob_for_sample'
  in EnbPI.fit() to avoid ndarray/list type conflict
- Split test_conformal.py into test_conformal.py (no sklearn needed)
  and test_enbpi.py (requires sklearn) so conformal_interval tests
  run in CI even without sklearn installed
- Fix ruff-format on test files
